refactor(daemon): split process policy tests
