+++ tasks/todo.md
@@ -65,11 +65,15 @@ priority application, Windows kill-on-close Job Object ownership, and the POSIX 
 primitive now use the neutral process facade. Product policy and diagnostics remain in their
 callers. Platform tests pass 32/32, focused daemon jobserver tests pass 2/2, and targeted
 production clippy passes with warnings denied. The exact-occurrence baseline fell from 336 to
-155 after also moving stdio redirection, crash-context interpretation, command window setup,
-CLI stack selection, and pre-spawn owner-death selection; its cache version is 0.1.6. Docker
+127 after also moving stdio redirection, crash-context interpretation, command window setup,
+CLI stack selection, and pre-spawn owner-death selection; its cache version is 0.1.7. Docker
 Desktop's Linux engine pipe was unavailable, so the
 Linux Dylint runtime gate remains required before publication; local baseline wiring validation
 passes.
+
+The two touched oversized production files were split without moving product policy:
+`process.rs` is 721 lines and `child_watchdog.rs` is 696 lines; their existing test modules
+now live in purpose-named sibling `_tests.rs` files.
 
 ## Phase 5 (create #1370 sub-issue) plan — platform::executable + platform::host
